Consolidate logging configuration.
