+++ kolibri/utils/options.py
@@ -14,6 +14,8 @@ except NotImplementedError:
     # This module can't work on this OS
     psutil = None
 
+
+from kolibri.utils.logger import get_base_logging_config
 from kolibri.plugins.utils.options import extend_config_spec
 
 
@@ -181,44 +183,9 @@ def get_logger(KOLIBRI_HOME):
     """
     We define a minimal default logger config here, since we can't yet load up Django settings.
     """
-    config = {
-        "version": 1,
-        "disable_existing_loggers": False,
-        "formatters": {
-            "simple_date_file": {
-                "()": "kolibri.core.logger.utils.formatter.KolibriLogFileFormatter",
-                "format": "%(levelname)s %(asctime)s %(module)s %(message)s",
-            },
-            "color": {
-                "()": "colorlog.ColoredFormatter",
-                "format": "%(log_color)s%(levelname)-8s %(message)s",
-            },
-        },
-        "handlers": {
-            "console": {
-                "level": "INFO",
-                "class": "logging.StreamHandler",
-                "formatter": "color",
-            },
-            "file": {
-                "level": "INFO",
-                "class": "kolibri.core.logger.utils.handler.KolibriTimedRotatingFileHandler",
-                "filename": os.path.join(KOLIBRI_HOME, "logs", "kolibri.txt"),
-                "formatter": "simple_date_file",
-                "when": "midnight",
-                "backupCount": 30,
-            },
-        },
-        "loggers": {
-            "kolibri": {
-                "handlers": ["console", "file"],
-                "level": "INFO",
-                "propagate": False,
-            }
-        },
-    }
+    from kolibri.utils.conf import LOG_ROOT
 
-    logging.config.dictConfig(config)
+    logging.config.dictConfig(get_base_logging_config(LOG_ROOT))
     return logging.getLogger(__name__)
